requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==3.0.1
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
executing==2.0.1
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.13.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.2.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_resources==6.4.5
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pillow==10.2.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.2.0
psutil==6.0.0
pure-eval==0.2.2
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
requests==2.32.3
rich==13.9.2
ruff==0.6.9
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.6
sympy==1.12
tbb==2021.11.0
tomlkit==0.12.0
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchinfo==1.8.0
torchvision==0.18.1+cu121
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
wcwidth==0.2.13
websockets==12.0
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.14.0
